test_2.3/requirements.txt
Flask
SQLAlchemy
psycopg2-binary
Pydantic
ffmpeg-python
flask-cors
